requirements.txt
paramiko==2.0.1
botocore==1.5.14
awsebcli==3.9.1
boto3==1.4.4